# openapi.yaml
openapi: 3.0.3
info:
  title: API Tienda Simple
  description: API REST con operaciones CRUD completas para gestionar clientes, productos y pedidos.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Servidor local de desarrollo
...
paths:
...
  /clientes:
    get:
      tags:
        - Clientes
      summary: Obtener todos los clientes
      responses:
        '200':
          description: Lista de clientes obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
              example:
                - id_cliente: 1
                  nombre: "Ana García"
                  email: "ana.garcia@email.com"
                  telefono: "555-1001"
                  created_at: "2026-06-25T09:00:00Z"
                - id_cliente: 2
                  nombre: "Luis Pérez"
                  email: "luis.perez@email.com"
                  telefono: "555-1002"
                  created_at: "2026-06-25T09:05:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Clientes
      summary: Crear un nuevo cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
            example:
              nombre: "María López"
              email: "maria.lopez@email.com"
              telefono: null
      responses:
        '201':
          description: Cliente creado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
              example:
                id_cliente: 3
                nombre: "María López"
                email: "maria.lopez@email.com"
                telefono: null
                created_at: "2026-06-25T09:10:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /clientes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID único del cliente
        schema:
          type: integer
    get:
      tags:
        - Clientes
      summary: Obtener un cliente por su ID
      responses:
        '200':
          description: Cliente encontrado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
              example:
                id_cliente: 1
                nombre: "Ana García"
                email: "ana.garcia@email.com"
                telefono: "555-1001"
                created_at: "2026-06-25T09:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Clientes
      summary: Actualizar un cliente existente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
            example:
              nombre: "Ana García de Mendoza"
              email: "ana.garcia.nueva@email.com"
              telefono: "555-9999"
      responses:
        '200':
          description: Cliente actualizado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
              example:
                id_cliente: 1
                nombre: "Ana García de Mendoza"
                email: "ana.garcia.nueva@email.com"
                telefono: "555-9999"
                created_at: "2026-06-25T09:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Clientes
      summary: Eliminar un cliente (Elimina en cascada sus pedidos)
      responses:
        '204':
          description: Cliente eliminado correctamente (No Content)
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ----------------------------------------------------
  # ENDPOINTS: PRODUCTOS
  # ----------------------------------------------------
  /productos:
    get:
      tags:
        - Productos
      summary: Obtener todos los productos
      responses:
        '200':
          description: Lista de productos obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Producto'
              example:
                - id_producto: 1
                  nombre: "Laptop Pro 15"
                  descripcion: "Laptop de alto rendimiento, 16 GB RAM"
                  precio: 1299.99
                  stock: 10
                - id_producto: 2
                  nombre: "Mouse Inalámbrico"
                  descripcion: "Mouse ergonómico con receptor USB"
                  precio: 29.99
                  stock: 50
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Productos
      summary: Crear un nuevo producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoInput'
            example:
              nombre: "Teclado Mecánico"
              descripcion: "Teclado retroiluminado RGB"
              precio: 89.99
              stock: 25
      responses:
        '201':
          description: Producto creado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producto'
              example:
                id_producto: 3
                nombre: "Teclado Mecánico"
                descripcion: "Teclado retroiluminado RGB"
                precio: 89.99
                stock: 25
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /productos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID único del producto
        schema:
          type: integer
    get:
      tags:
        - Productos
      summary: Obtener un producto por su ID
      responses:
        '200':
          description: Producto encontrado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producto'
              example:
                id_producto: 1
                nombre: "Laptop Pro 15"
                descripcion: "Laptop de alto rendimiento, 16 GB RAM"
                precio: 1299.99
                stock: 10
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Productos
      summary: Actualizar un producto existente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoInput'
            example:
              nombre: "Laptop Pro 15 V2"
              descripcion: "Laptop de alto rendimiento, 32 GB RAM"
              precio: 1499.99
              stock: 8
      responses:
        '200':
          description: Producto actualizado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producto'
              example:
                id_producto: 1
                nombre: "Laptop Pro 15 V2"
                descripcion: "Laptop de alto rendimiento, 32 GB RAM"
                precio: 1499.99
                stock: 8
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Productos
      summary: Eliminar un producto (Restringido si tiene pedidos asociados)
      responses:
        '204':
          description: Producto eliminado correctamente (No Content)
        '400':
          description: Error de solicitud o violación de clave foránea (Restricted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: "No se puede eliminar el producto porque tiene pedidos activos vinculados."
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ----------------------------------------------------
  # ENDPOINTS: PEDIDOS
  # ----------------------------------------------------
  /pedidos:
    get:
      tags:
        - Pedidos
      summary: Obtener todos los pedidos
      responses:
        '200':
          description: Lista de pedidos obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pedido'
              example:
                - id_pedido: 1
                  id_cliente: 1
                  id_producto: 1
                  cantidad: 1
                  total: 1299.99
                  fecha: "2026-06-25T09:15:00Z"
                - id_pedido: 2
                  id_cliente: 1
                  id_producto: 2
                  cantidad: 2
                  total: 59.98
                  fecha: "2026-06-25T09:20:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Pedidos
      summary: Registrar un nuevo pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoInput'
            example:
              id_cliente: 2
              id_producto: 3
              cantidad: 1
              total: 89.99
      responses:
        '201':
          description: Pedido registrado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
              example:
                id_pedido: 3
                id_cliente: 2
                id_producto: 3
                cantidad: 1
                total: 89.99
                fecha: "2026-06-25T09:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pedidos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID único del pedido
        schema:
          type: integer
    get:
      tags:
        - Pedidos
      summary: Obtener un pedido por su ID
      responses:
        '200':
          description: Pedido encontrado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
              example:
                id_pedido: 1
                id_cliente: 1
                id_producto: 1
                cantidad: 1
                total: 1299.99
                fecha: "2026-06-25T09:15:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Pedidos
      summary: Actualizar un pedido existente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoInput'
            example:
              id_cliente: 1
              id_producto: 1
              cantidad: 2
              total: 2599.98
      responses:
        '200':
          description: Pedido modificado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
              example:
                id_pedido: 1
                id_cliente: 1
                id_producto: 1
                cantidad: 2
                total: 2599.98
                fecha: "2026-06-25T09:15:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Pedidos
      summary: Eliminar/Cancelar un pedido
      responses:
        '204':
          description: Pedido eliminado correctamente (No Content)
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
...
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Código identificador del error (generalmente mapea el HTTP status)
        message:
          type: string
          description: Explicación detallada y legible del error generado

    # Esquemas: Clientes
    Cliente:
      type: object
      required:
        - id_cliente
        - nombre
        - email
        - created_at
      properties:
        id_cliente:
          type: integer
          format: int32
        nombre:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 150
        telefono:
          type: string
          maxLength: 20
          nullable: true
        created_at:
          type: string
          format: date-time
    ClienteInput:
      type: object
      required:
        - nombre
        - email
      properties:
        nombre:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 150
        telefono:
          type: string
          maxLength: 20
          nullable: true

    # Esquemas: Productos
    Producto:
      type: object
      required:
        - id_producto
        - nombre
        - precio
        - stock
      properties:
        id_producto:
          type: integer
          format: int32
        nombre:
          type: string
          maxLength: 150
        descripcion:
          type: string
          nullable: true
        precio:
          type: number
          format: float
        stock:
          type: integer
          default: 0
    ProductoInput:
      type: object
      required:
        - nombre
        - precio
      properties:
        nombre:
          type: string
          maxLength: 150
        descripcion:
          type: string
          nullable: true
        precio:
          type: number
          format: float
        stock:
          type: integer
          default: 0

    # Esquemas: Pedidos
    Pedido:
      type: object
      required:
        - id_pedido
        - id_cliente
        - id_producto
        - cantidad
        - total
        - fecha
      properties:
        id_pedido:
          type: integer
          format: int32
        id_cliente:
          type: integer
          format: int32
          description: ID referenciado de la tabla clientes
        id_producto:
          type: integer
          format: int32
          description: ID referenciado de la tabla productos
        cantidad:
          type: integer
          default: 1
        total:
          type: number
          format: float
        fecha:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Solicitud incorrecta o datos mal formados
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: "La sintaxis de la solicitud es inválida o faltan campos obligatorios."
    NotFound:
      description: El recurso solicitado no fue encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: "El recurso solicitado con el ID provisto no existe."
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 500
            message: "Ocurrió un error inesperado en el servidor al procesar la solicitud."